Error solucionado en reporte de cierre de caja y agregado a la base de datos el exento de iva.

+++ FacturacionElectronica/CLS/resumen.cs
@@ -8,25 +8,25 @@ namespace FacturacionElectronica.CLS
 {
     public class resumen
     {
-        public int totalNoSuj { get; set; }
-        public int totalExenta { get; set; }
+        public Double totalNoSuj { get; set; }
+        public Double totalExenta { get; set; }
         public Double totalGravada { get; set; }
         public Double subTotalVentas { get; set; }
-        public int descuNoSuj { get; set; }
-        public int descuExenta { get; set; }
+        public Double descuNoSuj { get; set; }
+        public Double descuExenta { get; set; }
         public int descuGravada { get; set; }
         public int porcentajeDescuento { get; set; }
-        public int totalDescu { get; set; }
+        public Double totalDescu { get; set; }
         public object tributos { get; set; }
         public Double subTotal { get; set; }
-        public int ivaRete1 { get; set; }
-        public int reteRenta { get; set; }
+        public Double ivaRete1 { get; set; }
+        public Double reteRenta { get; set; }
         public Double montoTotalOperacion { get; set; }
-        public int totalNoGravado { get; set; }
+        public Double totalNoGravado { get; set; }
         public Double totalPagar { get; set; }
         public string totalLetras { get; set; }
-        public int totalIva { get; set; }
-        public int saldoFavor { get; set; }
+        public Double totalIva { get; set; }
+        public Double saldoFavor { get; set; }
         public int condicionOperacion { get; set; }
         public List<pagos> pagos { get; set; }
         public object numPagoElectronico { get; set; }
